Move pip commands into their own module
