+++ crates/uv/src/commands/pip/freeze.rs
