+++ crates/uv/src/commands/pip/mod.rs
@@ -1,0 +1,8 @@
+pub(crate) mod check;
+pub(crate) mod compile;
+pub(crate) mod freeze;
+pub(crate) mod install;
+pub(crate) mod list;
+pub(crate) mod show;
+pub(crate) mod sync;
+pub(crate) mod uninstall;
