+++ crates/uv/src/commands/pip/show.rs
