+++ crates/uv/src/commands/pip/sync.rs
