+++ crates/uv/src/commands/pip/uninstall.rs
